Refactor to follow embedders directory structure
