+++ haystack_experimental/components/converters/image/document_to_image.py
