+++ haystack_experimental/components/converters/image/file_to_image.py
