+++ haystack_experimental/components/converters/image/image_utils.py
